Atualizacao automática

+++ webApp/src/main/java/com/super_bits/Controle_Usuario_Basico/paginas/PgAcessos.java
@@ -104,7 +104,7 @@ public class PgAcessos extends MB_PaginaConversation implements ItfPgAcessos {
     }
 
     @Override
-    public int getId() {
+    public Long getId() {
         return 0;
     }
